requirements.txt
aiogram==3.7.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
aiofiles==23.2.1
openai-whisper==20231117
torch==2.1.0
openai-whisper
static-ffmpeg
python-dotenv